requirements.txt
chattymarkov==1.3.0

# Discord
discord.py==2.0.0
typing_extensions
rapptz-asqlite==1.0.1
git+https://github.com/Rapptz/discord-ext-menus
